+++ src/pcs/kzg10/structs.rs
@@ -22,13 +22,13 @@ use ark_ec::AffineRepr;
 pub struct KZG10Commitment<E: Pairing> {
     /// the actual commitment is an affine point.
     pub com: E::G1Affine,
-    pub nv: usize,
+    pub nv: u8,
 }
 impl<E: Pairing> PolynomialCommitment<E::ScalarField> for KZG10Commitment<E> {
-    fn log_size(&self) -> usize {
+    fn log_size(&self) -> u8 {
         self.nv
     }
-    fn set_log_size(&mut self, nv: usize) {
+    fn set_log_size(&mut self, nv: u8) {
         self.nv = nv;
     }
 }
